+++ src/components/projectDetail/ApplyModal/ApplyModal.tsx
@@ -7,9 +7,9 @@ import { positionList } from 'utils/positions';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { updateApplicants, updateAppliedProject } from 'apis/postDetail';
 import { findWithCollectionName } from 'apis/findWithCollectionName';
-import ApplyButtonArea from './ApplyButtonArea';
+import ApplyModalButtonArea from './ApplyModalButtonArea';
 import ApplyMotiveArea from './ApplyMotiveArea';
-import ApplyPositionArea from './ApplyPositonArea';
+import ApplyPositionArea from './ApplyPositionArea';
 import ValidationToastPopup from 'components/common/ValidationToastPopup';
 import COLORS from 'assets/styles/colors';
 import MobileAlert from 'components/common/mobile/MobileAlert';
@@ -131,7 +131,7 @@ const ApplyModal = ({ isOpen, message, onClickEvent, pid }: props) => {
               />
             </MobileContentContainer>
             {/* 아니오, 지원하기 버튼 */}
-            <ApplyButtonArea
+            <ApplyModalButtonArea
               userData={userData}
               motive={motive}
               setMotive={setMotive}
@@ -172,7 +172,7 @@ const ApplyModal = ({ isOpen, message, onClickEvent, pid }: props) => {
           <ApplyMotiveArea motive={motive} setMotive={setMotive} />
         </WebContentContainer>
         {/* 아니오, 지원하기 버튼 */}
-        <ApplyButtonArea
+        <ApplyModalButtonArea
           userData={userData}
           motive={motive}
           setMotive={setMotive}
